+++ HTTP/Server.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <string>
 #include <unordered_map>
-#include <cstdint>
 #include <functional>
 #include <winsock2.h>
 #include <ws2tcpip.h>
@@ -19,7 +18,7 @@ namespace HTTP
 	
     struct Response
     {
-        uint16_t StatusCode = 0;
+        int StatusCode = 0;
         std::string Status = "";
         std::string Version = "";
         std::unordered_map<std::string, std::string> Headers{};
@@ -38,7 +37,7 @@ namespace HTTP
         Response& Build();
 	    ResponseBuilder& Reset();
         ResponseBuilder& Version(const std::string& Version); 
-	    ResponseBuilder& StatusCode(uint16_t StatusCode);
+	    ResponseBuilder& StatusCode(int StatusCode);
 	    ResponseBuilder& Status(const std::string& Status);
 	    ResponseBuilder& Body(const std::string& Body);
 	    ResponseBuilder& Header(const std::string& Key, const std::string& Value);
@@ -47,18 +46,18 @@ namespace HTTP
     class Server
     {
     private:
-	    static constexpr uint8_t MaxConnections = 32;
-	    static constexpr uint16_t BufferSize = 32768;
-	    static constexpr uint16_t ReadBufferSize = 8192;
+	    static constexpr int MaxConnections = 32;
+	    static constexpr int BufferSize = 32768;
+	    static constexpr int ReadBufferSize = 8192;
 
-	    uint8_t ConnectionCount = 0;
+	    int ConnectionCount = 0;
         pollfd* ConnectionList = nullptr;
         char* MessageBuffer = nullptr;
 	    std::unordered_map<std::string, std::function<Response(const Request&)>> Routes{};
 			
 	    void HandleNewConnection();
-	    void HandleClientData(uint8_t& Index);
-	    void SendResponse(int32_t ClientFD, const Response& Res);
+	    void HandleClientData(int& Index);
+	    void SendResponse(int ClientFD, const Response& Res);
         
         std::string CreateResponseString(const Response& Res);
 	    Request ParseRequest(const std::string& RawData);
